docs: toastgroup을 toast에 통합 (#218)
